requirements.txt
asgiref==3.4.1
boto3==1.18.3
botocore==1.21.3
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-sqlprint-middleware==0.1.3
django-storages==1.11.1
djangorestframework==3.12.4
idna==3.2
jmespath==0.10.0
psycopg2==2.9.1
pyfcm==1.5.3
Pygments==2.9.0
pygments-pprint-sql==0.1.0
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
stripe==2.60.0
typing-extensions==3.10.0.0
urllib3==1.26.6
